Moved tests in another folders.
